Create game.html
I didn't come up with this 100% on my own, this was from a book and I followed along with it.  However, I did find a weird bug where the apple can appear in the snake body, and I came up with a fix for that so I'm pretty proud of it lol.  The game is functional, but I didn't get around to making things pretty, also didn't get around to implementing a #lives feature.  I did this mainly to help me learn, not to showcase an actual finished product.